Added initial code for graceful shutdown (Issue #45)
This change adds the initial code for supporting graceful
shutdown of the services. It is still incomplete. Eventually
the shutdown_signal() function will be moved to a separate file
so it can be used for all services (currently api and ingestor).
Moreover, we need to understand exactly what shutdown means for
web services in this context so that catching a signal actually
results in desirable shutdown actions.