Used POST instead of GET for /product endpoint
The reason for this is that it potentially becomes impractical, or even
impossible, to encode the input JSON in the query string. For example,
a request to /product for a product type that supports arrays (typically
of station numbers) in its input can potentially fail with 414 URI Too
Long. The cURL syntax for POST can in fact be simpler than for GET since
we don't have to escape special chars in the query string:

curl -X POST 'localhost:3000/product?product_type=SineWave' \
    -d '{"from_time":1,"to_time":20,"time_resolution":1, \
    "min_value":-10,"max_value":10,"frequency":0.05}'

POST and GET each have their pros and cons, but POST seems to be best
in this case.