Renamed pool->db_pool + minimized passing of db_pool
This change does two things: First, pool is renamed to db_pool
for consistency with other code. Second, the pool is kept only
by the product types that need it, thus saving us from having to
pass it around in general functions. This makes the code clearer
by each product type specifying which storage pools it needs once
and for all in the new() function called on initialization of the
product type registry.