ingestion/legacy/checked: downgrade missing sensor to warning
Kvalobs sends a lot of these (~2 per day) and they seem malformed. I don't want
to get an alert for every one of these, So I've modified the parser to permit
them and emit a warning instead. I've also asked in the kvalobs-drift chat if
they have some other recommendation for handling these.